requirements.txt
discord.py
flask
rustplus
asyncio
gunicorn